regrouped material classes into different files since material was already 1000 lines long
