+++ demos/demo_mirrors.py
@@ -30,7 +30,8 @@ from distutils.version import StrictVersion
 
 
 from core import raster
-from core import material
+from core.material_isotropic import ConstantIndexGlass
+from core.material_anisotropic import AnisotropicMaterial
 from core import surfShape
 from core.optical_element import OpticalElement
 from core.optical_element_analysis import OpticalElementAnalysis
@@ -59,8 +60,8 @@ s = OpticalSystem()
 
 lc0 = s.addLocalCoordinateSystem(LocalCoordinates(name="object", decz=0.0), refname=s.rootcoordinatesystem.name)
 
-#air = material.AnisotropicMaterial(lc0, myeps)  # tests for anisotropic mirror
-air = material.ConstantIndexGlass(lc0, 1.0)
+#air = AnisotropicMaterial(lc0, myeps)  # tests for anisotropic mirror
+air = ConstantIndexGlass(lc0, 1.0)
 s.material_background = air
 
 lc1 = s.addLocalCoordinateSystem(LocalCoordinates(name="m1", decz=50.0, tiltx=-math.pi/8), refname=lc0.name) # objectDist
